Bug 1335651 - Setup an index path in the gecko.cache namespace for toolchain builds. r=dustin
This allows to find them and optimize them out during the taskgraph
optimization phase, and will allow to get toolchain artifacts through a
mach command for developers.

The index path is generated similarly to git trees or mercurial
manifests, and allows to find the right task corresponding to the the
contents of the files in the task `extra.resources` along the toolchain
scripts.

`when.files-changed` is not used when a task has index paths because we
need tasks to happen independently of whether there were changes to
those files when the index or artifacts expire.

UltraBlame original commit: e7e02e3c2e56edb9ba8bff03716052ae261ebb26