Bug 1340451. Don't use weak frames for reflow callbacks in nsMenuFrame. r=mats
Weak frames are stored in a linked list, so are not fast. Hence they are not meant to be used when there can be a lot of them active at the same time. When constructing large selects with lots of options we create one for every option.

There's no need to use a weak frame for a reflow callback as we can just cancel the reflow callback in the Destroy function.

UltraBlame original commit: 6af6e3b0247f89ab98d04f16f294c6b21b9730cc