Bug 1345109 - Ensure |mach python-test| errors out if no tests are found, r=mshal
Because test_objects was a generator, using it in the condition always returned True,
even if no tests were found. But extending test_objects to the manifest, converts it
to a list. So this patch simply moves the 'no tests' check a bit later on.

MozReview-Commit-ID: JpETWD1WQWH

UltraBlame original commit: 233d22402d7a65be5e9ac62141189ed70d182b88