Bug 1333714 - Allow highlighter to be hidden even if current node is not valid;r=zer0
When a highlighter was created for a node that has been removed, it can no longer be
hidden due to a check performed in the hide method of the basic auto-refresh class.

While it makes sense not to display a highlighter for an invalid node, hiding a
highlighter should always remain possible.

MozReview-Commit-ID: ChkmecJeqy9

UltraBlame original commit: e3a0d990896d176a21fa7c9e9f6e32b7f191e3de