Bug 1347569: don't add optmized tasks to task.dependencies; r=Callek
`task.dependencies` is part of the internal taskgraph calculations, so it must
only refer to tasks that are being created, and not to those which already
exist.  `task.task['dependencies']`, on the other hand, is what is given to
`queue.createTask` and expresses all dependencies.

MozReview-Commit-ID: GJ6JVj6JMBz

UltraBlame original commit: 742130b3c1249ba4001d25698c2db380a9e09013