Bug 1349114 - label CallPrintCallback for HTMLCanvasElement. r=smaug
Since canvas element is only run on main thread, and OwnerDoc is never null,
we should be able to use the new Dispatch API directly.

A trivial indent fix is included in this patch as well.
Since this is a one-line-change, I decided not to split this to a separated
patch.

MozReview-Commit-ID: ExjUdOdPGd4

UltraBlame original commit: b1f3d98c2533e8341cd48375f4dd7f6732b3a906