Bug 1348820 - Always collect metadata from visited pages. r=Grisha
We are going to need this in the future and starting collection even before releasing
Activity Stream will create a better experience once we turn it on.

And this flag is hard to miss. So let's just get rid of it.

MozReview-Commit-ID: 5oDzXhpQdSA

UltraBlame original commit: 2fffcd4ce3e439eeb5f88cc524b69a78704f51fd