Bug 1344748 - Skip recommended Marionette prefs for Firefox tests; r=maja_zf,whimboo
The recommended Marionette preferences for use in automation are not so
useful when testing the browser itself.  Many of the tests for Firefox
have highly specific requirements, and setting the recommended Marionette
preferences could for these mean a deterioration of test coverage.

The majority of test harnesses utilising Marionette use it to install
unsigned add-ons at runtime, and these preferences are not relevant in
this context.  The preferences they use are defined in prefs_general.js.

To avoid an unfortunate and unintended regression in test coverage,
we skip using the recommended Marionette preferences when we test the
browser itself.

MozReview-Commit-ID: 5h3cGPfbicV

UltraBlame original commit: 1598e2db1cf6bc8df116cff6a90ff382820801b8