Bug 1347425 - Part 1: Move site-specific Navigator.userAgent overrides to a more efficient C++ implementation; r=baku,schien
Necko computes this information in order to set the User-Agent header in
the parent process.  This header is set on all outgoing requests, and
therefore in the content process we can easily copy the value of this
header from the document channel object instead of the inefficient sync
IPC that currently happens in order to access the JavaScript
implementation living in the parent process.

UltraBlame original commit: 9206126bc23284918193314e0ef0692afe8b16c5