Bug 1350887 - Include Marionette prefs amongst defaults; r=ted,whimboo
The Marionette component ships in Firefox, but is not enabled by default.
We want to facilitate activating Marionette at runtime by flipping
the marionette.enabled preference, and showing the Marionette related
preferences in about:config helps discoverability.

It is also useful to rely on the preferences' default values so that
they do not have to be hardcoded in the component.

When Marionette is enabled by setting marionette.enabled to true, a set of
recommended automation preferences found in testing/marionette/server.js
are set if the user has not overriden/user-defined one of them and
marionette.prefs.recommended is true (default).  When Marionette is
stopped, the altered preferences are reset.

MozReview-Commit-ID: 3HLnEI0TEBB

UltraBlame original commit: 40cd9ca1553eae8e051a0208a122d7f3eadeda80