Bug 1345434 - Implementation of a virtual canvas technique for grid highlighter; r=pbro
A virtual canvas is basically a canvas that seems bigger than is actually is.
The technique consists in moving a fixed sized canvas during the scrolling, when
is needed, to give the illusion that it always covers the entire document.

MozReview-Commit-ID: Hp4cUZaBdm8

UltraBlame original commit: 194a927196496822ac2cf471a929ca72e84f3367