Bug 1351677 - When restoring a window, swap the initially selected tab with the desired selected tab instead of tab switching. r=mikedeboer
When restoring a window, it's cheaper if we move the initially selected tab to the
index of the tab that's supposed to be selected in the restored state, rather than
switching to that tab.

If it turns out that moving that tab is not possible (if, for example, the user
context IDs of the two tabs to swap don't match, since the userContextIds are
set at tab construction time), then we fall back to tab switching.

MozReview-Commit-ID: L3qP40K5DaJ

UltraBlame original commit: cb2a92a861321c8c5c630ce1964d55d4a869f6fd