Bug 1355570 - Simplify nsCSSRenderingGradients r=mattwoodrow
In nsCSSRenderingGradients the logic for handling many error conditions
and normalizing the gradient stops is shared between the paint path and
the WebRender path. WebRender now normalizes gradients and handles any
error conditions that it needs to. There are some conditions that are not
problems for it, like repeating radial gradients with stops below zero.

This commit undoes the work done in bug 1341101 to share this logic. Some
conditions were moved around in bug 1341101 to make things simpler, and
that has been undone. Now the paint path is identical to how it was
originally.

There is one exception, which is ResolveMidpoints which is kept between
both code paths. This should be safe.

MozReview-Commit-ID: LMhMNXNquXM

UltraBlame original commit: 30005fc4bb990f6ca0d97445494654eea51f1ac5