Bug 1085393 - Remove pre-Runtime ADB support. r=ochameau
Removes support for old versions of ADB Helper which haven't been in use for
years.

It also removes `test_telemetry` because it referenced this code and is also
disabled for all configurations in a product we intend to remove.

MozReview-Commit-ID: AUAbwxVczWb

UltraBlame original commit: 031bad6381d7adf67948d130f9ce753a045c5e8e